[CORRECTION][REGLES DE GESTION] Supprime la notion 'SUPPRIME_REPONSE' des règles de gestion
Cette notion a été introduite suite à une incompréhension. Elle avait pour but de supprimer les réponses apportées (non applicable) dans le cas des systèmes industriels si
l'Aidant modifiait le contexte en répondant oui pour le périmètre des systèmes industriels.
Cela ne fait pas sens, on ne modifie pas les réponses apportées dans ce cas là.
