+++ mon-aide-cyber-api/test/diagnostic/Diagnostic.spec.ts
@@ -20,7 +20,6 @@ import {
   genereLaRestitution,
   Indicateurs,
   MesurePriorisee,
-  ReponseDonnee,
 } from '../../src/diagnostic/Diagnostic';
 import { uneAssociation } from '../constructeurs/constructeurAssociation';
 import { Question, ReponsePossible } from '../../src/diagnostic/Referentiel';
@@ -43,7 +42,7 @@ describe('Diagnostic', () => {
   const questions = uneListeDe7QuestionsToutesAssociees();
 
   describe('Réponses au diagnostic', () => {
-    const unDiagnosticAvecUneRegleDeGestionAjouteReponse = (
+    const unDiagnosticAvecUneRegleDeGestion = (
       premiereQuestion: Question,
       premiereReponsePossible: ReponsePossible,
       deuxiemeQuestion: Question,
@@ -57,7 +56,7 @@ describe('Diagnostic', () => {
                 .aChoixUnique('Question à tiroir ?')
                 .avecReponsesPossibles([
                   uneReponsePossible()
-                    .ajouteUneRegleAjouteReponses([
+                    .ajouteUneRegleDeGestion([
                       {
                         identifiantQuestion: premiereQuestion.identifiant,
                         reponseDonnee: premiereReponsePossible.identifiant,
@@ -77,7 +76,7 @@ describe('Diagnostic', () => {
         )
         .construis();
 
-    it('Applique la règle de gestion AJOUTE_REPONSE liée à la question', async () => {
+    it('Applique la règle de gestion liée à la question', async () => {
       const premiereReponsePossible = uneReponsePossible().construis();
       const premiereQuestion = uneQuestion()
         .avecReponsesPossibles([
@@ -92,7 +91,7 @@ describe('Diagnostic', () => {
           deuxiemeReponsePossible,
         ])
         .construis();
-      const diagnostic = unDiagnosticAvecUneRegleDeGestionAjouteReponse(
+      const diagnostic = unDiagnosticAvecUneRegleDeGestion(
         premiereQuestion,
         premiereReponsePossible,
         deuxiemeQuestion,
@@ -119,91 +118,6 @@ describe('Diagnostic', () => {
         diagnostic.referentiel['thematique-2'].questions[0].reponseDonnee
           .reponseUnique
       ).toStrictEqual(deuxiemeReponsePossible.identifiant);
-    });
-
-    const unDiagnosticAvecUneRegleDeGestionSupprimeReponse = (
-      premiereQuestion: Question,
-      deuxiemeQuestion: Question
-    ) =>
-      unDiagnostic()
-        .avecUnReferentiel(
-          unReferentiel()
-            .ajouteUneQuestionAuContexte(
-              uneQuestion()
-                .aChoixUnique('Question à tiroir ?')
-                .avecReponsesPossibles([
-                  uneReponsePossible()
-                    .ajouteUneRegleSupprimeReponses([
-                      premiereQuestion.identifiant,
-                      deuxiemeQuestion.identifiant,
-                    ])
-                    .construis(),
-                ])
-                .construis()
-            )
-            .ajouteUneThematique('thematique-1', [
-              uneQuestion()
-                .avecReponsesPossibles([uneReponsePossible().construis()])
-                .construis(),
-              premiereQuestion,
-            ])
-            .ajouteUneThematique('thematique-2', [deuxiemeQuestion])
-            .construis()
-        )
-        .construis();
-
-    it('Applique la règle de gestion SUPPRIME_REPONSE liée à la question', async () => {
-      const premiereReponsePossible = uneReponsePossible().construis();
-      const premiereQuestion = uneQuestion()
-        .avecReponsesPossibles([
-          premiereReponsePossible,
-          uneReponsePossible().construis(),
-        ])
-        .construis();
-      const deuxiemeReponsePossible = uneReponsePossible().construis();
-      const deuxiemeQuestion = uneQuestion()
-        .avecReponsesPossibles([
-          uneReponsePossible().construis(),
-          deuxiemeReponsePossible,
-        ])
-        .construis();
-      const diagnostic = unDiagnosticAvecUneRegleDeGestionSupprimeReponse(
-        premiereQuestion,
-        deuxiemeQuestion
-      );
-
-      ajouteLaReponseAuDiagnostic(
-        diagnostic,
-        unCorspsDeReponse()
-          .concernantLaQuestion(premiereQuestion)
-          .pourLaThematique('thematique-1')
-          .avecLaReponse(premiereQuestion.reponsesPossibles[0].identifiant)
-          .construis()
-      );
-      ajouteLaReponseAuDiagnostic(
-        diagnostic,
-        unCorspsDeReponse()
-          .concernantLaQuestion(diagnostic.referentiel.contexte.questions[0])
-          .pourLaThematique('contexte')
-          .avecLaReponse(
-            diagnostic.referentiel.contexte.questions[0].reponsesPossibles[0]
-              .identifiant
-          )
-          .construis()
-      );
-
-      expect(
-        diagnostic.referentiel['thematique-1'].questions[1].reponseDonnee
-      ).toStrictEqual<ReponseDonnee>({
-        reponseUnique: null,
-        reponsesMultiples: [],
-      });
-      expect(
-        diagnostic.referentiel['thematique-2'].questions[0].reponseDonnee
-      ).toStrictEqual<ReponseDonnee>({
-        reponseUnique: null,
-        reponsesMultiples: [],
-      });
     });
   });
